+++ skoolkit/skoolasm.py
@@ -1,4 +1,4 @@
-# Copyright 2008-2020 Richard Dymond (rjdymond@gmail.com)
+# Copyright 2008-2021 Richard Dymond (rjdymond@gmail.com)
 #
 # This file is part of SkoolKit.
 #
@@ -243,6 +243,11 @@ class AsmWriter:
 
     def expand_list(self, text, index):
         return self._ignore_block(text, index, LIST_MARKER, LIST_END_MARKER)
+
+    def expand_over(self, text, index):
+        if self.handle_unsupported_macros:
+            return skoolmacro.parse_over(text, index, self.fields)
+        raise skoolmacro.UnsupportedMacroError()
 
     def expand_plot(self, text, index):
         if self.handle_unsupported_macros: