+++ sphinx/source/changelog.rst
@@ -4,6 +4,7 @@ Changelog
 8.5b1
 -----
 * Dropped support for Python 3.5
+* Added the :ref:`OVER` macro (for superimposing one frame on another)
 * Added the ``Wrap`` configuration parameter for
   :ref:`sna2skool.py <sna2skool-conf>` (for controlling whether to disassemble
   an instruction that wraps around the 64K boundary)